refactor(ui): extract SettingsPage icons, config, nav, and standalone modals
